+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-Worldthreader 2.1.0 is the first release for Minecraft 1.21.5.
+Worldthreader 2.1.1 for Minecraft 1.21.5 fixes multiple thread safety issues.
 Please report any issues you encounter to the [issue tracker of worldthreader](https://github.com/2No2Name/worldthreader/issues). As worldthreader likely comes with massive mod compatibility issues, please do not report crashes and issues to other mods' issue trackers before confirming the issue without worldthreader.
 
 ## Fixes:
 
+- Fix thread safety issue of Brain Sensor entity targeting range
 - Fix theoretical thread safety issue of redstone torch burnout after switching singleplayer worlds
 
 ## Known issues